third_party/blink/web_tests/fast/js/console-non-string-values-expected.txt

CONSOLE MESSAGE: I am a string
CONSOLE MESSAGE: 42
Should see two console messages, the first a string, and the second a number.
PASS successfullyParsed is true

TEST COMPLETE
